backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.1
python-multipart>=0.0.9
tensorflow-cpu>=2.16.1
Pillow>=10.3.0
numpy>=1.26.4
h5py>=3.11.0
pydantic>=2.7.4
